+++ assets/rocket.js
@@ -3,32 +3,27 @@
         "dom": "google",
         "anchorsel": "h3.r a",
         "input": "input[name='q'][type='text']:first"
-    },
-    {
+    },{
         "dom": "thepiratebay",
         "anchorsel": ".detLink"
-    },
-    {
+    },{
         "dom": "quora",
         "anchorsel": ".question_link"
-    },
-    {
+    },{
         "dom": "cheatography",
         "anchorsel": "#body_shadow strong a"
-    },
-    {
+    },{
         "dom": "readthedocs",
         "anchorsel": "#id_search_result a"
-    },
-    {
+    },{
         "dom": "ynet",
         "anchorsel": ".smallheader , #mStrips div div .text12, .whtbigheader:nth-child(2), .blkbigheader span"
-    },
-    {
+    },{
         "dom": "nana10"
-    },
-    {
-        "dom"      : "stackoverflow|serverfault|superuser|meta.stackoverflow|askubuntu|stackapps|answers.onstartups|mathoverflow"
+       ,"textsel": "b"
+       ,"anchorsel": "a.GenenalHompageLinkNoBold"
+    },{
+        "dom"      : "\/^http://.*\\.stackexchange\\.com\/|stackoverflow|serverfault|superuser|meta.stackoverflow|askubuntu|stackapps|answers.onstartups|mathoverflow"
        ,"anchorsel": ".result-link a"
        ,"input"    : "[name='q']:last"
     },{