+++ CHANGELOG.md
@@ -5,6 +5,8 @@ v2.22.0
 - Rocket-framework: ADD support multiple domains -> one selector
 - Using underscore.js for more functional goodness
 - Ctrl+Shift+[1-9] Opens in a new tab AND switches to it.
+- Rocket-framework: REGEX.
+
 v2.21.0
 - expose a function pi6process ( page_object ) to dev console.